+++ TicTacToe/Move.cpp
@@ -1,9 +1,9 @@
 #include "Move.h"
 
-Move::Move(const uint8_t players_amount, const int v) : evaluation(std::vector<int>(players_amount, v)) {}
-Move::Move(const uint8_t y, const uint8_t x, const std::vector<int>& results) : has_move(true), y(y), x(x), evaluation(results) {}
+Move::Move(const int players_amount, const int v) : evaluation(std::vector<int>(players_amount, v)) {}
+Move::Move(const int y, const int x, const std::vector<int>& results) : has_move(true), y(y), x(x), evaluation(results) {}
 
-bool Move::isWorse(const Move& other, uint8_t current_player) const
+bool Move::isWorse(const Move& other, int current_player) const
 {
 	if (other.evaluation[current_player] > evaluation[current_player])
 	{
@@ -12,7 +12,7 @@ bool Move::isWorse(const Move& other, uint8_t current_player) const
 	return other.evaluation[current_player] > evaluation[current_player];
 }
 
-bool Move::isSame(const Move& other, uint8_t current_player) const
+bool Move::isSame(const Move& other, int current_player) const
 {
 	return other.evaluation[current_player] == evaluation[current_player] && moves_remaining == other.moves_remaining;
 }
