+++ TicTacToe/Move.h
@@ -4,15 +4,15 @@
 struct Move
 {
 	bool has_move = false;
-	uint8_t y;
-	uint8_t x;
-	uint8_t moves_remaining = 0;
+	int y;
+	int x;
+	int moves_remaining = 0;
 	std::vector<int> evaluation;
 
-	Move(const uint8_t players_amount, const int v);
-	Move(const uint8_t y, const uint8_t x, const std::vector<int>& results);
+	Move(const int players_amount, const int v);
+	Move(const int y, const int x, const std::vector<int>& results);
 
-	bool isWorse(const Move& other, uint8_t current_player) const;
+	bool isWorse(const Move& other, int current_player) const;
 
-	bool isSame(const Move& other, uint8_t current_player) const;
+	bool isSame(const Move& other, int current_player) const;
 };
